Moved database operations to the main project; search results page layout skeleton; back to search button.
