+++ .vs/ExtendedMemory/xs/UserPrefs.xml
@@ -1,19 +1,20 @@
 ﻿<Properties StartupConfiguration="{E9F5DE4C-CDA9-47EA-99D0-8CAAA9DAB05B}|Default">
-  <MonoDevelop.Ide.Workbench ActiveDocument="iOS/DataAccess/MemoryDatabase.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="ExtendedMemory/Views/SearchResultsPage.xaml">
     <Files>
-      <File FileName="iOS/DataAccess/MemoryDatabase.cs" Line="130" Column="29" />
-      <File FileName="ExtendedMemory/Views/SearchPage.xaml.cs" Line="80" Column="84" />
-      <File FileName="ExtendedMemory/Views/SavePage.xaml.cs" Line="2" Column="10" />
-      <File FileName="ExtendedMemory/Views/SearchResultsPage.xaml.cs" Line="16" Column="74" />
-      <File FileName="ExtendedMemory/DataAccess/IMemoryDatabase.cs" Line="15" Column="63" />
+      <File FileName="ExtendedMemory/DataAccess/MemoryDatabase.cs" Line="1" Column="3" />
+      <File FileName="ExtendedMemory/Views/SavePage.xaml.cs" Line="47" Column="62" />
+      <File FileName="ExtendedMemory/Views/SearchPage.xaml.cs" Line="28" Column="21" />
+      <File FileName="ExtendedMemory/Views/SearchResultsPage.xaml.cs" Line="13" Column="9" />
+      <File FileName="ExtendedMemory/Views/SavePage.xaml" Line="59" Column="17" />
+      <File FileName="ExtendedMemory/Views/SearchPage.xaml" Line="96" Column="15" />
+      <File FileName="ExtendedMemory/Views/SearchResultsPage.xaml" Line="29" Column="15" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug|iPhoneSimulator" />
   <MonoDevelop.Ide.DebuggingService.Breakpoints>
-    <BreakpointStore>
-      <Breakpoint file="/Users/rohit/Drive_Sync/Code/ExtendedMemory/ExtendedMemory/ExtendedMemory/Views/SearchResultsPage.xaml.cs" relfile="ExtendedMemory/Views/SearchResultsPage.xaml.cs" line="25" column="1" />
-    </BreakpointStore>
+    <BreakpointStore />
   </MonoDevelop.Ide.DebuggingService.Breakpoints>
+  <MonoDevelop.Ide.ItemProperties.ExtendedMemory.Droid AndroidDesignerPreferredTheme="Theme.DeviceDefault.NoActionBar.Fullscreen" />
   <MonoDevelop.Ide.DebuggingService.PinnedWatches />
   <MonoDevelop.Ide.ItemProperties.ExtendedMemory.iOS PreferredExecutionTarget="MonoDevelop.IPhone.IPhoneSimulatorTarget.3A0C59C0-B9C1-4D54-B47A-4BFA58120CB4" />
   <MultiItemStartupConfigurations />
