+++ ExtendedMemory/Views/SearchResultsPage.xaml
@@ -1,9 +1,29 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" x:Class="ExtendedMemory.Views.SearchResultsPage">
-    <ContentPage.Content>
-    </ContentPage.Content>
-    <Label Grid.Row="0" Grid.Column="0"
+<?xml version="1.0" encoding="utf-8"?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" 
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" 
+             x:Class="ExtendedMemory.Views.SearchResultsPage"
+             Title="Search Results" >
+    
+    <Grid RowSpacing="40">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="100" />
+            <RowDefinition Height="50" />
+            <RowDefinition Height="100" />
+        </Grid.RowDefinitions>
+
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+        
+
+        <Button Grid.Row="0" Grid.Column="0"
+            Text="Back" Clicked="BackToSearch" />
+            
+        <Label Grid.Row="1" Grid.Column="0" Grid.ColumnSpan="2"
                Text="Result"
                HorizontalTextAlignment="Center"
                VerticalTextAlignment="Center" />
+
+    </Grid>
 </ContentPage>
