+++ ExtendedMemory/Views/SearchResultsPage.xaml.cs
@@ -1,4 +1,5 @@
-﻿using ExtendedMemory.DataAccess;
+﻿using System;
+using ExtendedMemory.DataAccess;
 using ExtendedMemory.Models;
 using Xamarin.Forms;
 
@@ -13,16 +14,28 @@ namespace ExtendedMemory.Views
 
         public SearchResultsPage(SearchParams searchParams)
         {
-            var searchResult = DependencyService.Get<IMemoryDatabase>().Get(searchParams);
-            if (searchResult.IsSuccess)
-            {
-                DisplayAlert("Success", searchResult.Item.Count + " results found", "OK");
-            }
-            else {
-                DisplayAlert("Failure", "Something went wrong", "OK");
-            }
+            InitializeComponent();
 
-            Application.Current.MainPage = new HomePage() {CurrentPage = new SearchPage()};
+            var searchResult = new MemoryDatabase().Get(searchParams);
+            Device.BeginInvokeOnMainThread(() =>
+            {
+                if (searchResult.IsSuccess)
+                {
+                    DisplayAlert("Success", searchResult.Item.Count + " results found", "OK");
+                }
+                else
+                {
+                    DisplayAlert("Failure", "Something went wrong", "OK");
+                }
+            });
+        }
+
+        void BackToSearch(object sender, EventArgs args)
+        {
+            Application.Current.MainPage = new HomePage()
+            {
+                CurrentPage = new SearchPage()
+            };
         }
     }
 }
