+++ iOS/ExtendedMemory.iOS.csproj
@@ -166,7 +166,6 @@
     <Compile Include="AppDelegate.cs" />
     <Compile Include="Main.cs" />
     <Compile Include="Helpers\GetLocation_IOS.cs" />
-    <Compile Include="DataAccess\MemoryDatabase.cs" />
     <Compile Include="Helpers\FileLocation.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -179,7 +178,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Helpers\" />
-    <Folder Include="DataAccess\" />
   </ItemGroup>
   <Import Project="..\ExtendedMemory\ExtendedMemory.projitems" Label="Shared" Condition="Exists('..\ExtendedMemory\ExtendedMemory.projitems')" />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\iOS\Xamarin.iOS.CSharp.targets" />
